Let´s make the canvas work again at least in the server.
- The id doesn´t much anymore with the index of the array, so use the
game state instead of the game mode to get the player color and index

+++ Source/ColouringBook/ColouringBookInkDrop.h
@@ -39,7 +39,7 @@ public:
 	/** Returns ProjectileMovement subobject **/
 	FORCEINLINE UProjectileMovementComponent* GetProjectileMovement() const { return ProjectileMovement; }
 	/* Returns the number ID of the actor that spawned this actor */
-	FORCEINLINE uint8 GetOwnerID() const { return ownerID; }
+	FORCEINLINE int32 GetOwnerID() const { return ownerID; }
 
 	UFUNCTION()
 	void OnReplicatedOwnerID();
